+++ client/web/src/enterprise/insights/pages/dashboards/dashboard-page/components/dashboards-content/DashboardsContent.tsx
@@ -9,9 +9,10 @@ import { Button, Tooltip } from '@sourcegraph/wildcard'
 
 import { HeroPage } from '../../../../../../../components/HeroPage'
 import { LimitedAccessLabel } from '../../../../../components'
-import { InsightDashboard, isVirtualDashboard, ALL_INSIGHTS_DASHBOARD } from '../../../../../core'
+import { ALL_INSIGHTS_DASHBOARD } from '../../../../../constants'
+import { InsightDashboard, isVirtualDashboard } from '../../../../../core'
 import { useCopyURLHandler, useUiFeatures } from '../../../../../hooks'
-import { AddInsightModal } from '../add-insight-modal/AddInsightModal'
+import { AddInsightModal } from '../add-insight-modal'
 import { DashboardMenu, DashboardMenuAction } from '../dashboard-menu/DashboardMenu'
 import { DashboardSelect } from '../dashboard-select/DashboardSelect'
 import { DeleteDashboardModal } from '../delete-dashboard-modal/DeleteDashboardModal'
@@ -96,7 +97,7 @@ export const DashboardsContent: React.FunctionComponent<React.PropsWithChildren<
                 <span className={styles.dashboardSelectLabel}>Dashboard:</span>
 
                 <DashboardSelect
-                    value={currentDashboard?.id}
+                    dashboard={currentDashboard}
                     dashboards={dashboards}
                     className={classNames(styles.dashboardSelect, 'mr-2')}
                     onSelect={handleDashboardSelect}