+++ cmd/searcher/main.go
@@ -155,8 +155,9 @@ func run(logger log.Logger) error {
 	service := &search.Service{
 		Store: &search.Store{
 			FetchTar: func(ctx context.Context, repo api.RepoName, commit api.CommitID) (io.ReadCloser, error) {
-				// We pass in a nil sub-repo permissions checker here since searcher needs access
-				// to all data in the archive
+				// We pass in a nil sub-repo permissions checker and an internal actor here since
+				// searcher needs access to all data in the archive.
+				ctx = actor.WithInternalActor(ctx)
 				return git.ArchiveReader(ctx, nil, repo, gitserver.ArchiveOptions{
 					Treeish: string(commit),
 					Format:  gitserver.ArchiveFormatTar,
@@ -167,8 +168,9 @@ func run(logger log.Logger) error {
 				for i, p := range paths {
 					pathspecs[i] = gitdomain.PathspecLiteral(p)
 				}
-				// We pass in a nil sub-repo permissions checker here since searcher needs access
-				// to all data in the archive
+				// We pass in a nil sub-repo permissions checker and an internal actor here since
+				// searcher needs access to all data in the archive.
+				ctx = actor.WithInternalActor(ctx)
 				return git.ArchiveReader(ctx, nil, repo, gitserver.ArchiveOptions{
 					Treeish:   string(commit),
 					Format:    gitserver.ArchiveFormatTar,
@@ -183,7 +185,11 @@ func run(logger log.Logger) error {
 			DB:                 db,
 		},
 
-		GitDiffSymbols:      git.DiffSymbols,
+		GitDiffSymbols: func(ctx context.Context, repo api.RepoName, commitA, commitB api.CommitID) ([]byte, error) {
+			// As this is an internal service call, we need an internal actor.
+			ctx = actor.WithInternalActor(ctx)
+			return git.DiffSymbols(ctx, repo, commitA, commitB)
+		},
 		MaxTotalPathsLength: maxTotalPathsLength,
 
 		Log: logger,