+++ internal/types/types.go
@@ -1663,14 +1663,15 @@ type SearchContextRepositoryRevisions struct {
 
 type EncryptableSecret = encryption.Encryptable
 
-func NewEmptySecret() *EncryptableSecret {
-	return NewUnencryptedSecret("")
-}
-
+// NewUnencryptedSecret creates an EncryptableSecret that *may* be encrypted in
+// the future, but the current value has not yet been encrypted.
 func NewUnencryptedSecret(value string) *EncryptableSecret {
 	return encryption.NewUnencrypted(value)
 }
 
+// NewEncryptedSecret creates an EncryptableSecret that has come from an
+// encrypted source. In this case you need to provide the keyID and key in order
+// to be able to decrypt it.
 func NewEncryptedSecret(cipher, keyID string, key encryption.Key) *EncryptableSecret {
 	return encryption.NewEncrypted(cipher, keyID, key)
 }
@@ -1681,9 +1682,18 @@ type Webhook struct {
 	// The primary key, used for sorting and pagination
 	ID int32
 	// UUID is the ID we display externally and will appear in the webhook URL
-	UUID            uuid.UUID
-	CodeHostKind    string
-	CodeHostURN     string
+	UUID         uuid.UUID
+	CodeHostKind string
+	CodeHostURN  string
+	// Secret can be in one of three states:
+	//
+	// 1. nil, no secret provided.
+	// 2. Provided but not encrypted.
+	// 3. Provided and encrypted.
+	//
+	// For 2 and 3 you interact with it in the same way and just assume that it IS
+	// encrypted. All the methods on EncryptableSecret will just pass around the raw
+	// value and encryption / decryption methods are noops.
 	Secret          *EncryptableSecret
 	CreatedAt       time.Time
 	UpdatedAt       time.Time